requirements.txt
boto3==1.35.71
python-dotenv==1.0.1
pytest==8.3.4
pytest-mock==3.14.0
pytest-cov==6.0.0